Move methods from ApplicationRecord to GithubMethods concern
