+++ app/handlers/card_converted_handler.rb
@@ -1,6 +1,6 @@
 class CardConvertedHandler < BasicEventHandler
   def execute!
-    issue = Issue.find_or_create payload
+    issue = Issue.update_or_create payload
             .project_card.issue.to_hash.merge(user: sender)
     card = Card.find_by(github_id: payload.project_card.github_id)
     card.update(issue: issue)
