+++ app/handlers/card_edited_handler.rb
@@ -1,5 +1,5 @@
 class CardEditedHandler < BasicEventHandler
   def execute!
-    Card.find_and_update payload.project_card.to_hash
+    Card.update_or_create payload.project_card.to_hash
   end
 end
