+++ app/handlers/column_created_handler.rb
@@ -1,6 +1,6 @@
 class ColumnCreatedHandler < BasicEventHandler
   def execute!
-    project = Project.find_or_create payload.project_column.project
+    project = Project.update_or_create payload.project_column.project.to_hash
     Column.create payload.project_column.to_hash.merge(project: project)
   end
 end
