+++ app/handlers/issue_assigned_handler.rb
@@ -1,7 +1,7 @@
 class IssueAssignedHandler < BasicEventHandler
   def execute!
     issue = Issue.find_by(github_id: payload.issue.github_id)
-    assignee = User.find_or_create payload.assignee
+    assignee = User.update_or_create payload.assignee.to_hash
     issue.update(assignee: assignee)
   end
 end
