+++ app/handlers/project_created_handler.rb
@@ -1,5 +1,5 @@
 class ProjectCreatedHandler < BasicEventHandler
   def execute!
-    Project.find_or_create(payload.project.to_hash.merge(user: sender))
+    Project.update_or_create(payload.project.to_hash.merge(user: sender))
   end
 end
