+++ app/handlers/project_edited_handler.rb
@@ -1,5 +1,5 @@
 class ProjectEditedHandler < BasicEventHandler
   def execute!
-    Project.find_and_update payload.project.to_hash
+    Project.update_or_create payload.project.to_hash
   end
 end
