+++ app/handlers/pull_request_opened_handler.rb
@@ -1,6 +1,6 @@
 class PullRequestOpenedHandler < BasicEventHandler
   def execute!
-    PullRequest.find_or_create(
+    PullRequest.update_or_create(
       payload.pull_request.to_hash.merge(user: sender)
     )
   end
