+++ app/jobs/concerns/handle_event_job.rb
@@ -1,4 +1,6 @@
 module HandleEventJob
+  extend ActiveSupport::Concern
+
   def perform(data)
     payload = PayloadStruct.new data
     handler = get_handler(payload.action)
