+++ app/models/issue.rb
@@ -1,4 +1,6 @@
 class Issue < ApplicationRecord
+  include GithubMethods
+
   belongs_to :user
   has_one :card, dependent: :destroy
   belongs_to :assignee, class_name: 'User', optional: true
