+++ app/models/project.rb
@@ -1,4 +1,6 @@
 class Project < ApplicationRecord
+  include GithubMethods
+
   belongs_to :user, optional: true
   has_many :columns, dependent: :destroy
 end
