+++ app/models/pull_request.rb
@@ -1,3 +1,5 @@
 class PullRequest < ApplicationRecord
+  include GithubMethods
+
   belongs_to :user
 end
